event-store-project/pyproject.toml
# ────────────────────────────────────────────────────────────────────────────
#TASK 2: Create Requirements File
#────────────────────────────────────────────────────────────────────────────

[project]
name = "event-store-project"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"

dependencies = [
    "python-dotenv>=1.0.1",
    "psycopg2-binary>=2.9.9",
    "pytest>=7.4.3",
    "python-dateutil>=2.8.2",
]

[build-system]
requires = ["setuptools>=78.1.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
